generator/pyproject.toml
[project]
name = "olam-ha-ai-generator"
version = "0.1.0"
description = "Automated post generator for the 'עולם ה AI' Hebrew blog"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "firecrawl-py>=1.0.0",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "anyascii>=0.3.2",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
